FE-1239 - move type url pull to CreateForm
