+++ src/pages/domains/CreatePage.js
@@ -1,29 +1,10 @@
 import React from 'react';
-import { usePageFilters } from 'src/hooks';
 import { Page } from 'src/components/matchbox';
 import { PageLink } from 'src/components/links';
 import { BASE_URL } from './constants';
 import Domains from './components';
 
-const initFilters = {
-  type: { defaultValue: 'sending' },
-};
-
 export default function CreatePage() {
-  const { filters } = usePageFilters(initFilters);
-  const { type } = filters;
-
-  function getDomainCreateType() {
-    switch (type) {
-      case 'sending':
-      case 'bounce':
-      case 'tracking':
-        return type;
-      default:
-        return;
-    }
-  }
-
   return (
     <Domains.Container>
       <Page
@@ -34,7 +15,7 @@ export default function CreatePage() {
         }}
         title="Add a Domain"
       >
-        <Domains.CreateForm defaultType={getDomainCreateType()} />
+        <Domains.CreateForm />
       </Page>
     </Domains.Container>
   );
